+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed types for constraints, tainted and errors when using intersections and unions in schemas.
+- Fixed SuperDebug collapsed height by preventing global css leak.
 
 ## [2.8.1] - 2024-03-07